Split the unit test for density fitting RSCF and USCF into 4 tasks.
